feat: consolidate CLI and add structured logging
Major architectural improvements:
- Consolidated internal/cli/ (8 files) into single internal/cli.go
- Moved business logic to domain packages (daemon.LaunchWithClaude, daemon.ShowLastLines, daemon.FollowLogs)
- Added structured logging with slog and log rotation via lumberjack
- Made 'athena' launch daemon + Claude Code by default
- Made 'athena logs' tail by default (--follow=true)

Breaking changes:
- Removed JSON config format support (YAML-only)
- Removed wrapper scripts from release workflow
- Changed default behavior: 'athena' now runs 'athena code' functionality

Configuration improvements:
- Added logging flags (--log-format, --log-file)
- Updated all docs to reference YAML-only config
- Simplified config precedence: CLI → YAML → env → defaults

Release workflow improvements:
- Removed 200+ lines of wrapper script generation
- Simplified installation with integrated daemon management
- Updated release documentation and examples

Architecture benefits:
- Single source of truth for CLI commands
- Clear separation: CLI orchestrates, domains implement
- Better testability with domain functions
- Follows single-file organization standard

All tests passing (109 tests across 6 packages)
Security scan clean with govulncheck
