--- athena.example.json
@@ -1,9 +1,0 @@
-{
-  "port": "11434",
-  "api_key": "your-openrouter-api-key-here",
-  "base_url": "https://openrouter.ai/api",
-  "model": "google/gemini-2.0-flash-exp:free",
-  "opus_model": "anthropic/claude-3-opus",
-  "sonnet_model": "anthropic/claude-3.5-sonnet",
-  "haiku_model": "anthropic/claude-3.5-haiku"
-}
