+++ go.mod
@@ -2,7 +2,10 @@ module athena
 
 go 1.24.7
 
-require github.com/spf13/cobra v1.10.1
+require (
+	github.com/goccy/go-yaml v1.18.0
+	github.com/spf13/cobra v1.10.1
+)
 
 require (
 	github.com/inconshreveable/mousetrap v1.1.0 // indirect
