+++ internal/server/server.go
@@ -4,8 +4,9 @@ import (
 	"bytes"
 	"encoding/json"
 	"io"
-	"log"
+	"log/slog"
 	"net/http"
+	"strings"
 	"time"
 
 	"athena/internal/config"
@@ -24,10 +25,11 @@ func New(cfg *config.Config) *Server {
 
 // Start starts the HTTP server
 func (s *Server) Start() error {
+
 	http.HandleFunc("/v1/messages", s.handleMessages)
 	http.HandleFunc("/health", s.handleHealth)
 
-	log.Printf("Starting server on port %s", s.cfg.Port)
+	slog.Info("starting server", "port", s.cfg.Port)
 
 	// Create server with proper timeouts for security
 	server := &http.Server{
@@ -44,11 +46,14 @@ func (s *Server) Start() error {
 func (s *Server) handleHealth(w http.ResponseWriter, _ *http.Request) {
 	w.WriteHeader(http.StatusOK)
 	if err := json.NewEncoder(w).Encode(map[string]string{"status": "ok"}); err != nil {
-		log.Printf("Failed to encode health response: %v", err)
+		slog.Error("failed to encode health response", "error", err)
 	}
 }
 
 func (s *Server) handleMessages(w http.ResponseWriter, r *http.Request) {
+	start := time.Now()
+	ctx := r.Context()
+
 	if r.Method != "POST" {
 		http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
 		return
@@ -67,8 +72,28 @@ func (s *Server) handleMessages(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
+	slog.Info("request received",
+		"method", "POST",
+		"path", "/v1/messages",
+		"model", req.Model,
+		"stream", req.Stream,
+	)
+
 	// Transform to OpenAI format
 	openAIReq := transform.AnthropicToOpenAI(req, s.cfg)
+
+	// Log provider routing if configured
+	providerInfo := "default"
+	if openAIReq.Provider != nil && len(openAIReq.Provider.Order) > 0 {
+		providerInfo = strings.Join(openAIReq.Provider.Order, ",")
+	}
+
+	slog.Info("routing request",
+		"from_model", req.Model,
+		"to_model", openAIReq.Model,
+		"provider", providerInfo,
+	)
+
 	openAIBody, err := json.Marshal(openAIReq)
 	if err != nil {
 		http.Error(w, "Failed to marshal OpenAI request", http.StatusInternalServerError)
@@ -78,7 +103,7 @@ func (s *Server) handleMessages(w http.ResponseWriter, r *http.Request) {
 	// Forward to OpenRouter
 	client := &http.Client{}
 	url := s.cfg.BaseURL + "/v1/chat/completions"
-	openRouterReq, err := http.NewRequest("POST", url, bytes.NewReader(openAIBody))
+	openRouterReq, err := http.NewRequestWithContext(ctx, "POST", url, bytes.NewReader(openAIBody))
 	if err != nil {
 		http.Error(w, "Failed to create request", http.StatusInternalServerError)
 		return
@@ -100,6 +125,20 @@ func (s *Server) handleMessages(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 	defer resp.Body.Close()
+
+	duration := time.Since(start)
+
+	// Extract actual provider from OpenRouter response headers
+	actualProvider := resp.Header.Get("X-OpenRouter-Provider")
+	if actualProvider == "" {
+		actualProvider = "unknown"
+	}
+
+	slog.Info("response received",
+		"status", resp.StatusCode,
+		"duration_ms", duration.Milliseconds(),
+		"actual_provider", actualProvider,
+	)
 
 	// Handle response based on streaming
 	if req.Stream {
